+++ app/src/androidTest/java/androidTest/ExampleInstrumentedTest.java
@@ -1,4 +1,4 @@
-package com.example.bbqbuddy;
+package androidTest;
 
 import android.content.Context;
