+++ app/src/main/java/backend/DatabaseController.java
@@ -1,0 +1,4 @@
+package backend;
+
+public class DatabaseController {
+}
