+++ app/src/main/java/ui/MainActivity.java
@@ -1,8 +1,10 @@
-package com.example.bbqbuddy;
+package ui;
 
 import android.os.Bundle;
 
 import androidx.appcompat.app.AppCompatActivity;
+
+import com.example.bbqbuddy.R;
 
 public class MainActivity extends AppCompatActivity {
